move kitchen to common
