+++ src/main/java/org/labs/common/Kitchen.java
@@ -1,4 +1,4 @@
-package org.labs.orderedlocks;
+package org.labs.common;
 
 import java.util.concurrent.atomic.AtomicInteger;
 import lombok.extern.slf4j.Slf4j;
